powernv: Properly panic on MCE
Hi surelock folk,

This relates to SW317149.

I'm still trying to determine the correct behaviour
in the MCE in Skiboot case. It's a bit more complex than I thought:
the kernel MCE handler should determine that it's unrecoverable, it's
just xmon that will think the address is in userspace. The kernel should
then OOPS but will attempt to stagger on.

It's not entirely clear how the kernel should behave in this
situation - whether it should explicitly panic or if the current behaviour
of OOPS+reboot is actually better. It seems that the rest of the code was
written expecting the kernel to panic, although that's not strictly true.
I'd like to understand _why_ we don't automatically panic before I propose
making us automatically panic: I'll ask around the office tomorrow.

In the mean time, could you please apply this, run _without_
xmon, and see if it's possible to reach the hang before we
panic out? Thanks!

Signed-off-by: Daniel Axtens <dja@axtens.net>