cxlflash: Fix to free irqs before cancelling work
An interrupt can come in after cancelling work from
the workq, and the processing of that interrupt
can cause another workq element to be queued up.
This can result in the original problem of
accessing free memory from the workq handler.

To avoid this condition it is better to free the
interrupts first, before cancelling work.