cxlflash: Correct usage of scnprintf in loop
The use of scnprintf() in cxlflash_show_port_lun_table() is incorrect
and can lead to a condition where the buffer is accessed beyond it's
size. This can lead to data corruption.

The issue is due to scnprintf() being called in a loop that modifies
the starting address of where to write data but doesn't take into
account the remaning space left in the buffer.

To remedy, decrease the amount of space already consumed from the total
length of the buffer each time scnprintf() is invoked. Additionally, to
be consistent with other drivers in how they approach using scnprintf()
in a loop, advance the buffer by the number of bytes already written
instead of using a hardcoded value.

Signed-off-by: Matthew R. Ochs <mrochs@linux.vnet.ibm.com>
Signed-off-by: Manoj N. Kumar <manoj@linux.vnet.ibm.com>