cxlflash: Fix to avoid potential deadlock on EEH
The SCSI WRITE_SAME16 can take quite a while to complete. Should an EEH occur
while in scsi_execute(), the EEH handler will attempt to recover. As part of
the recovery, the handler drains all currently running ioctls, waiting until
they have completed before proceeding with a reset. As this routine is used
on the ioctl path, this can create a condition where the EEH handler becomes
stuck, infinitely waiting for this ioctl thread. To avoid this behavior,
temporarily unmark this thread as an ioctl thread by releasing the ioctl read
semaphore. This will allow the EEH handler to proceed with a recovery while
this thread is still running. Once the scsi_execute() returns, reacquire the
ioctl read semaphore and check the adapter state in case it changed while
inside of scsi_execute(). The state check will wait if the adapter is still
being recovered or return a failure if the recovery failed. In the event that
the adapter reset failed, simply return the failure as the ioctl would be
unable to continue.

Reported-by: Brian King <brking@linux.vnet.ibm.com>
Signed-off-by: Matthew R. Ochs <mrochs@linux.vnet.ibm.com>
Signed-off-by: Manoj N. Kumar <manoj@linux.vnet.ibm.com>