cxlflash: Fix to resolve oops in wait_port_offline
If an asynchronous error interrupt is generated the processing cannot
be performed in the interrupt context. So the work is scheduled to an
asynchronous worker thread. Between the time the interrupt is generated
and the asynchronous worker thread runs, the MMIO space for the master
context may by unmapped either with an escalation of the reset process,
or due to EEH error recovery.

The oops could have a signature like this:

[  606.806583] NIP [c0000000000e20f8] kthread_data+0x28/0x40
[  606.806633] LR [c0000000000d99d0] wq_worker_sleeping+0x30/0x100
[  606.806694] Call Trace:
[  606.806721] [c000002ff2e835a0] [0000000000000050] 0x50 (unreliable)
[  606.806796] [c000002ff2e835d0] [c0000000000d99d0] wq_worker_sleeping+0x30/0x100
[  606.806884] [c000002ff2e83610] [c000000000a5feec] __schedule+0x69c/0x8a0
[  606.806959] [c000002ff2e83690] [c000000000a60134] schedule+0x44/0xc0
[  606.807034] [c000002ff2e836c0] [c0000000000b8aa0] do_exit+0x770/0xb90
[  606.807109] [c000002ff2e83790] [c0000000000209e0] die+0x300/0x460
[  606.807185] [c000002ff2e83820] [c00000000004f878] bad_page_fault+0xd8/0x150
[  606.807259] [c000002ff2e83890] [c000000000008680] handle_page_fault+0x2c/0x30
[  606.807338] --- interrupt: 300 at wait_port_offline.constprop.12+0x60/0x130 [cxlflash]

To resolve this issue the tearing down of the master context's MMIO
space should be gated by any asynchronous work item to completed,
This solution re-purposes the ioctl semaphore to be used for
any operation, including this asynchronous worker thread.